+++ packages/passkeys/passkeys_ios/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.4.0
+
 * Added `getAvailability` api to check for passkey support
+* Fixes issue with iOS 16.0 and above where the `DeviceNotSupported` exception was thrown when the device was supported
 
 ## 2.3.0