+++ packages/passkeys/passkeys_platform_interface/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.2.0
 * Added `hasPasskeySupport` to `AvailabilityType`
+* Separate `AvailabilityType` for each supported platform
 
 ## 2.1.0
 * Added getAvailability().